add setting to choose if the docks shall be opened at the original position or above the last active dock

+++ src/commonfrontend/core/ContentDockWidget.cpp
@@ -27,12 +27,13 @@
  * this class automatically updates the window title when AbstractAspect::caption() is changed
  * and holds the connection to the actual data visualized in this window via the pointer to \c AbstractPart.
  */
-ContentDockWidget::ContentDockWidget(AbstractPart* part)
+ContentDockWidget::ContentDockWidget(AbstractPart* part, bool deleteOnClose)
 	: ads::CDockWidget(part->name())
-	, m_part(part) {
+	, m_part(part)
+	, mDeleteOnClose(deleteOnClose) {
 	setWindowIcon(m_part->icon());
 	setWidget(m_part->view());
-	setFeature(DockWidgetDeleteOnClose, true);
+	setFeature(DockWidgetDeleteOnClose, deleteOnClose);
 	handleAspectDescriptionChanged(m_part);
 
 	// resize the MDI sub window to fit the content of the view
@@ -69,7 +70,8 @@ void ContentDockWidget::handleAspectAboutToBeRemoved(const AbstractAspect* aspec
 }
 
 void ContentDockWidget::closeEvent() {
-	m_part->deleteView();
+	if (mDeleteOnClose)
+		m_part->deleteView();
 }
 
 void ContentDockWidget::slotWindowStateChanged(Qt::WindowStates /*oldState*/, Qt::WindowStates newState) {
